move topology struct and functions to topology module
