+++ prince/include/parser.h
@@ -5,6 +5,8 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "topology.h"
+
 /* bc degree structures definition*/
 typedef struct _id_degree_bc{
 	char* id;
@@ -20,25 +22,6 @@ typedef struct _map_id_degree_bc{
 } map_id_degree_bc;
 
 
-typedef struct topology * topology_t;
-
-typedef struct node * node_t;
-
-/* topology structures definition*/
-struct topology {
-	int id_lenght;
-	char *protocol;
-	char *self_id;
-	struct node *first;
-};
-
-struct node {
-	char *id;
-	struct neighbor *neighbor_list;
-	node_t next;
-	struct local_address *addresses;
-};
-
 struct neighbor {
 	node_t id;
 	float weight;
@@ -51,16 +34,10 @@ struct local_address {
 	struct local_address *next;
 };
 
-int  add_neigh(topology_t topo, const char *source, const char *id, const double weight, int validity);
-int  add_node(topology_t topo, const char *id);
-
 struct neighbor* find_neigh(node_t source, node_t target);
-node_t           find_node(topology_t topo, const char *id);
 
 void free_bc_degree_map(map_id_degree_bc * map);
 
-topology_t new_topo(int type);
-void       free_topo(topology_t topo);
 topology_t parse_jsoninfo(char *buffer);
 topology_t parse_netjson(char* buffer);
